CLJS-647: Use js-obj function with non-string keys
When all args passed in key position to core.clj/js-obj are
strings, the macro efficiently expands the form to a JS object
literal. If not all keys are args, the slower path through
core.cljs/js-obj and Google Closure is taken, which evaluates
the keys and coerces them to strings.